requirements.txt
pytils
ipython
fabric
markdown
django-compressor
django-annoying